+++ packages/cli-core/src/package-manager-controller/npm_package_manager_controller.ts
@@ -4,6 +4,7 @@ import { execa as _execa } from 'execa';
 import * as _path from 'path';
 import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_debugger_logger.js';
 import { PackageManagerControllerBase } from './package_manager_controller_base.js';
+import { NpmLockFileReader } from './lock-file-reader/npm_lock_file_reader.js';
 
 /**
  * NpmPackageManagerController is an abstraction around npm commands that are needed to initialize a project and install dependencies
@@ -18,13 +19,15 @@ export class NpmPackageManagerController extends PackageManagerControllerBase {
     protected readonly path = _path,
     protected readonly execa = _execa,
     protected readonly executeWithDebugLogger = _executeWithDebugLogger,
-    protected readonly existsSync = _existsSync
+    protected readonly existsSync = _existsSync,
+    protected readonly lockFileReader = new NpmLockFileReader()
   ) {
     super(
       cwd,
       'npm',
       ['init', '--yes'],
       'install',
+      lockFileReader,
       fsp,
       path,
       execa,