+++ packages/sandbox/src/file_watching_sandbox.test.ts
@@ -36,6 +36,7 @@ import {
   SSMClient,
   SSMServiceException,
 } from '@aws-sdk/client-ssm';
+import { EOL } from 'os';
 
 // Watcher mocks
 const unsubscribeMockFn = mock.fn();
@@ -196,7 +197,11 @@ void describe('Sandbox to check if region is bootstrapped', () => {
     assert.strictEqual(printer.log.mock.callCount(), 1);
     assert.strictEqual(
       printer.log.mock.calls[0].arguments[0],
-      'The given region has not been bootstrapped. Sign in to console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.'
+      `The region ${format.highlight(
+        region
+      )} has not been bootstrapped. Sign in to the AWS console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.${EOL}If this is not the region you are expecting to bootstrap, check for any AWS environment variables that may be set in your shell or use ${format.command(
+        '--profile <profile-name>'
+      )} to specify a profile with the correct region.`
     );
     assert.strictEqual(printer.log.mock.calls[0].arguments[1], undefined);
   });
@@ -227,7 +232,11 @@ void describe('Sandbox to check if region is bootstrapped', () => {
     assert.strictEqual(printer.log.mock.callCount(), 3);
     assert.strictEqual(
       printer.log.mock.calls[0].arguments[0],
-      'The given region has not been bootstrapped. Sign in to console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.'
+      `The region ${format.highlight(
+        region
+      )} has not been bootstrapped. Sign in to the AWS console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.${EOL}If this is not the region you are expecting to bootstrap, check for any AWS environment variables that may be set in your shell or use ${format.command(
+        '--profile <profile-name>'
+      )} to specify a profile with the correct region.`
     );
     assert.strictEqual(printer.log.mock.calls[0].arguments[1], undefined);
     assert.strictEqual(