+++ packages/sandbox/src/file_watching_sandbox.ts
@@ -37,6 +37,7 @@ import {
   BackendIdentifierConversions,
 } from '@aws-amplify/platform-core';
 import { LambdaFunctionLogStreamer } from './lambda_function_log_streamer.js';
+import { EOL } from 'os';
 
 /**
  * CDK stores bootstrap version in parameter store. Example parameter name looks like /cdk-bootstrap/<qualifier>/version.
@@ -121,12 +122,16 @@ export class FileWatchingSandbox extends EventEmitter implements Sandbox {
 
     this.filesChangesTracker = await createFilesChangesTracker(watchDir);
     const bootstrapped = await this.isBootstrapped();
+    // get region from an available sdk client;
+    const region = await this.ssmClient.config.region();
     if (!bootstrapped) {
       this.printer.log(
-        'The given region has not been bootstrapped. Sign in to console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.'
+        `The region ${format.highlight(
+          region
+        )} has not been bootstrapped. Sign in to the AWS console as a Root user or Admin to complete the bootstrap process, then restart the sandbox.${EOL}If this is not the region you are expecting to bootstrap, check for any AWS environment variables that may be set in your shell or use ${format.command(
+          '--profile <profile-name>'
+        )} to specify a profile with the correct region.`
       );
-      // get region from an available sdk client;
-      const region = await this.ssmClient.config.region();
       const bootstrapUrl = getBootstrapUrl(region);
       try {
         await this.open(bootstrapUrl);