+++ packages/auth-construct/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/auth-construct
 
+## 1.6.0
+
+### Minor Changes
+
+- ff2f2ce: add sms configuration and custom sms sender to auth
+
 ## 1.5.1
 
 ### Patch Changes