+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @aws-amplify/backend-cli
 
+## 1.4.8
+
+### Patch Changes
+
+- ff2f2ce: fixed violations to cause propagation lint rule
+- Updated dependencies [ff2f2ce]
+- Updated dependencies [ff2f2ce]
+  - @aws-amplify/backend-deployer@1.1.15
+  - @aws-amplify/schema-generator@1.2.7
+  - @aws-amplify/cli-core@1.2.3
+  - @aws-amplify/sandbox@1.2.10
+
+## 1.4.7
+
+### Patch Changes
+
+- 2dab201: Report cdk versions
+- Updated dependencies [2dab201]
+  - @aws-amplify/backend-deployer@1.1.14
+  - @aws-amplify/cli-core@1.2.2
+  - @aws-amplify/platform-core@1.6.0
+  - @aws-amplify/plugin-types@1.8.0
+
+## 1.4.6
+
+### Patch Changes
+
+- a7506f9: wraps no outputs found error from backend output client
+- Updated dependencies [a7506f9]
+- Updated dependencies [a7506f9]
+- Updated dependencies [a7506f9]
+- Updated dependencies [a7506f9]
+  - @aws-amplify/model-generator@1.0.12
+  - @aws-amplify/client-config@1.5.5
+  - @aws-amplify/sandbox@1.2.9
+  - @aws-amplify/backend-deployer@1.1.13
+  - @aws-amplify/platform-core@1.5.0
+  - @aws-amplify/plugin-types@1.7.0
+
+## 1.4.5
+
+### Patch Changes
+
+- 0ef53e2: add validation for sandbox set command secret-name
+- Updated dependencies [107600b]
+  - @aws-amplify/model-generator@1.0.11
+
 ## 1.4.4
 
 ### Patch Changes