+++ packages/platform-core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @aws-amplify/platform-core
 
+## 1.6.0
+
+### Minor Changes
+
+- 2dab201: Report cdk versions
+
+### Patch Changes
+
+- Updated dependencies [2dab201]
+  - @aws-amplify/plugin-types@1.8.0
+
+## 1.5.1
+
+### Patch Changes
+
+- a712983: Base64 encode serialized Amplify Errors
+
+## 1.5.0
+
+### Minor Changes
+
+- a7506f9: added data logging api to defineData
+
+### Patch Changes
+
+- a7506f9: add InsufficientMemorySpaceError wrapping
+- Updated dependencies [a7506f9]
+  - @aws-amplify/plugin-types@1.7.0
+
 ## 1.4.0
 
 ### Minor Changes