+++ packages/plugin-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/plugin-types
 
+## 1.8.0
+
+### Minor Changes
+
+- 2dab201: Report cdk versions
+
+## 1.7.0
+
+### Minor Changes
+
+- a7506f9: added data logging api to defineData
+
 ## 1.6.0
 
 ### Minor Changes